requirements.txt
flask==3.0.3
prometheus_client==0.21.0
pyJWT==2.10.1
pytest==8.4.2